+++ membership/management/commands/paper_reminders.py
@@ -2,12 +2,10 @@
 from __future__ import with_statement
 
 import logging
-
 from optparse import make_option
 
-from membership import bulk_utils
-
 from django.core.management.base import BaseCommand, CommandError
+from membership.billing import pdf_utils
 
 logger = logging.getLogger("paper_bills")
 
@@ -23,7 +21,7 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         try:
-            pdffile = bulk_utils.generate_pdf_reminders(memberid=options['member'])
+            pdffile = pdf_utils.generate_pdf_reminders(memberid=options['member'])
             if pdffile:
                 print "pdf file created: %s" % pdffile
             else:
