+++ membership/models.py
@@ -3,8 +3,11 @@
 from datetime import datetime, timedelta
 from decimal import Decimal
 import logging
+
+from membership.billing import pdf
 from membership.reference_numbers import barcode_4, group_right,\
     generate_membership_bill_reference_number
+
 logger = logging.getLogger("membership.models")
 import traceback
 
@@ -12,7 +15,8 @@ from cStringIO import StringIO
 
 from django.core.exceptions import ObjectDoesNotExist
 from django.db import models
-from django.db.models import Q, Sum
+from django.db.models import Q, Sum, Count
+
 from django.utils.translation import ugettext_lazy as _
 from django.conf import settings
 from django.template.loader import render_to_string
@@ -27,7 +31,6 @@ from utils import log_change, tupletuple_to_dict
 from email_utils import send_as_email, send_preapprove_email, send_duplicate_payment_notice
 from email_utils import bill_sender, preapprove_email_sender, duplicate_payment_sender
 
-from membership import pdf
 
 class BillingEmailNotFound(Exception): pass
 class MembershipOperationError(Exception): pass
@@ -627,6 +630,58 @@ class BillingCycle(models.Model):
         fees = Fee.objects.filter(for_this_type, not_before_start)
         vat_percentage = fees.latest('start').vat_percentage
         return vat_percentage
+
+    @classmethod
+    def get_reminder_billingcycles(cls, memberid=None):
+        """
+        Get queryset for BillingCycles with missing payments and witch have 2 or more bills already sent.
+        :param memberid:
+        :return:
+        """
+        if not settings.ENABLE_REMINDERS:
+            return cls.objects.none()
+
+        qs = cls.objects
+
+        # Single membership case
+        if memberid:
+            logger.info('memberid: %s' % memberid)
+            qs = qs.filter(membership__id=memberid)
+            qs = qs.exclude(bill__type='P')
+            return qs
+
+        # For all memberships in Approved state
+        qs = qs.annotate(bills=Count('bill'))
+        qs = qs.filter(bills__gt=2,
+                       is_paid__exact=False,
+                       membership__status='A',
+                       membership__id__gt=-1)
+        qs = qs.exclude(bill__type='P')
+        qs = qs.order_by('start')
+
+        return qs
+
+    @classmethod
+    def create_paper_remainder_list(cls, memberid=None):
+        """
+        Create list of BillingCycles with missing payments and which already don't have paper bill.
+        :param memberid: optional member id
+        :return: list of billingcycles
+        """
+        datalist = []
+        for cycle in cls.get_reminder_billingcycles(memberid).all():
+            # check if paper reminder already sent
+            cont = False
+            for bill in cycle.bill_set.all():
+                if bill.type == 'P':
+                    cont=True
+                    break
+            if cont:
+                continue
+
+            datalist.append(cycle)
+        return datalist
+
 
     def __unicode__(self):
         return str(self.start.date()) + "--" + str(self.end.date())
